test: fix resolve-target tests to avoid mock interference
- Move integration tests with mock.module() to test/isolated/ directory
  to run separately from unit tests (Bun mocks affect global module state)
- Keep pure function property tests in test/lib/resolve-target.test.ts
- Add test:isolated script for running isolated tests
- Coverage improved from 15.28% to 72.49% for resolve-target.ts

The isolated tests cover:
- resolveOrg: CLI flags, defaults, DSN detection, error handling
- resolveFromDsn: cache hits, API fetches, fallback behavior
- resolveOrgAndProject: flag validation, defaults, DSN, directory inference
- resolveAllTargets: monorepo multi-DSN, deduplication, empty results
