+++ package.json
@@ -19,6 +19,7 @@
     "lint:fix": "bunx ultracite fix",
     "test": "bun test",
     "test:unit": "bun test test/lib test/commands test/types",
+    "test:isolated": "bun test test/isolated",
     "test:e2e": "bun test test/e2e",
     "generate:skill": "bun run script/generate-skill.ts",
     "check:skill": "bun run script/check-skill.ts"
